moved code inside src/
https://github.com/nextauthjs/next-auth/issues/5268
did however find what caused the issue for me: I created an api folder in the root of my project. While it wasn't causing any issue locally, it was causing issues on Vercel, because apparently that is where the actual API pages go once the functions are deployed. So to avoid any conflicts with Vercel architecture, I moved my api folder to a src folder, along with pages and everything else that is the source code.